+++ node/src/test/kotlin/net/corda/node/services/vault/NodeVaultServiceTest.kt
@@ -403,7 +403,7 @@ class NodeVaultServiceTest : TestDependencyInjectionBase() {
             val freshKey = services.legalIdentityKey
 
             // Issue a txn to Send us some Money
-            val usefulBuilder = TransactionType.General.Builder(null).apply {
+            val usefulBuilder = TransactionBuilder(null).apply {
                 Cash().generateIssue(this, 100.DOLLARS `issued by` MEGA_CORP.ref(1), AnonymousParty(freshKey), DUMMY_NOTARY)
             }
             val usefulTX = megaCorpServices.signInitialTransaction(usefulBuilder)
@@ -416,7 +416,7 @@ class NodeVaultServiceTest : TestDependencyInjectionBase() {
             assertEquals(3, vaultSvc.getTransactionNotes(usefulTX.id).count())
 
             // Issue more Money (GBP)
-            val anotherBuilder = TransactionType.General.Builder(null).apply {
+            val anotherBuilder = TransactionBuilder(null).apply {
                 Cash().generateIssue(this, 200.POUNDS `issued by` MEGA_CORP.ref(1), AnonymousParty(freshKey), DUMMY_NOTARY)
             }
             val anotherTX = megaCorpServices.signInitialTransaction(anotherBuilder)
@@ -458,7 +458,7 @@ class NodeVaultServiceTest : TestDependencyInjectionBase() {
         val amount = Amount(1000, Issued(BOC.ref(1), GBP))
 
         // Issue then move some cash
-        val issueTx = TransactionBuilder(TransactionType.General, services.myInfo.legalIdentity).apply {
+        val issueTx = TransactionBuilder(services.myInfo.legalIdentity).apply {
             Cash().generateIssue(this,
                     amount, anonymousIdentity.party, services.myInfo.legalIdentity)
         }.toWireTransaction()
@@ -468,7 +468,7 @@ class NodeVaultServiceTest : TestDependencyInjectionBase() {
         val expectedIssueUpdate = Vault.Update(emptySet(), setOf(cashState), null)
 
         database.transaction {
-            val moveTx = TransactionBuilder(TransactionType.General, services.myInfo.legalIdentity).apply {
+            val moveTx = TransactionBuilder(services.myInfo.legalIdentity).apply {
                 service.generateSpend(this, Amount(1000, GBP), thirdPartyIdentity)
             }.toWireTransaction()
             service.notify(moveTx)