+++ node/src/test/kotlin/net/corda/node/services/vault/VaultWithCashTest.kt
@@ -10,6 +10,7 @@ import net.corda.core.node.services.VaultService
 import net.corda.core.node.services.consumedStates
 import net.corda.core.node.services.unconsumedStates
 import net.corda.core.transactions.SignedTransaction
+import net.corda.core.transactions.TransactionBuilder
 import net.corda.node.services.database.HibernateConfiguration
 import net.corda.node.services.schema.NodeSchemaService
 import net.corda.node.utilities.CordaPersistence
@@ -93,7 +94,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
         database.transaction {
             // A tx that sends us money.
             val freshKey = services.keyManagementService.freshKey()
-            val usefulBuilder = TransactionType.General.Builder(null)
+            val usefulBuilder = TransactionBuilder(null)
             Cash().generateIssue(usefulBuilder, 100.DOLLARS `issued by` MEGA_CORP.ref(1), AnonymousParty(freshKey), DUMMY_NOTARY)
             val usefulTX = megaCorpServices.signInitialTransaction(usefulBuilder)
 
@@ -101,7 +102,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             services.recordTransactions(usefulTX)
 
             // A tx that spends our money.
-            val spendTXBuilder = TransactionType.General.Builder(DUMMY_NOTARY)
+            val spendTXBuilder = TransactionBuilder(DUMMY_NOTARY)
             vault.generateSpend(spendTXBuilder, 80.DOLLARS, BOB)
             val spendPTX = services.signInitialTransaction(spendTXBuilder, freshKey)
             val spendTX = notaryServices.addSignature(spendPTX)
@@ -109,7 +110,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             assertEquals(100.DOLLARS, services.getCashBalance(USD))
 
             // A tx that doesn't send us anything.
-            val irrelevantBuilder = TransactionType.General.Builder(DUMMY_NOTARY)
+            val irrelevantBuilder = TransactionBuilder(DUMMY_NOTARY)
             Cash().generateIssue(irrelevantBuilder, 100.DOLLARS `issued by` MEGA_CORP.ref(1), BOB, DUMMY_NOTARY)
 
             val irrelevantPTX = megaCorpServices.signInitialTransaction(irrelevantBuilder)
@@ -147,7 +148,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
         backgroundExecutor.submit {
             database.transaction {
                 try {
-                    val txn1Builder = TransactionType.General.Builder(DUMMY_NOTARY)
+                    val txn1Builder = TransactionBuilder(DUMMY_NOTARY)
                     vault.generateSpend(txn1Builder, 60.DOLLARS, BOB)
                     val ptxn1 = notaryServices.signInitialTransaction(txn1Builder)
                     val txn1 = services.addSignature(ptxn1, freshKey)
@@ -177,7 +178,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
         backgroundExecutor.submit {
             database.transaction {
                 try {
-                    val txn2Builder = TransactionType.General.Builder(DUMMY_NOTARY)
+                    val txn2Builder = TransactionBuilder(DUMMY_NOTARY)
                     vault.generateSpend(txn2Builder, 80.DOLLARS, BOB)
                     val ptxn2 = notaryServices.signInitialTransaction(txn2Builder)
                     val txn2 = services.addSignature(ptxn2, freshKey)
@@ -219,7 +220,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             val linearId = UniqueIdentifier()
 
             // Issue a linear state
-            val dummyIssueBuilder = TransactionType.General.Builder(notary = DUMMY_NOTARY).apply {
+            val dummyIssueBuilder = TransactionBuilder(notary = DUMMY_NOTARY).apply {
                 addOutputState(DummyLinearContract.State(linearId = linearId, participants = listOf(freshIdentity)))
                 addOutputState(DummyLinearContract.State(linearId = linearId, participants = listOf(freshIdentity)))
             }
@@ -240,7 +241,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             val linearId = UniqueIdentifier()
 
             // Issue a linear state
-            val dummyIssueBuilder = TransactionType.General.Builder(notary = DUMMY_NOTARY)
+            val dummyIssueBuilder = TransactionBuilder(notary = DUMMY_NOTARY)
             dummyIssueBuilder.addOutputState(DummyLinearContract.State(linearId = linearId, participants = listOf(freshIdentity)))
             val dummyIssuePtx = notaryServices.signInitialTransaction(dummyIssueBuilder)
             val dummyIssue = services.addSignature(dummyIssuePtx)
@@ -251,7 +252,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             assertThat(vault.unconsumedStates<DummyLinearContract.State>()).hasSize(1)
 
             // Move the same state
-            val dummyMoveBuilder = TransactionType.General.Builder(notary = DUMMY_NOTARY).apply {
+            val dummyMoveBuilder = TransactionBuilder(notary = DUMMY_NOTARY).apply {
                 addOutputState(DummyLinearContract.State(linearId = linearId, participants = listOf(freshIdentity)))
                 addInputState(dummyIssue.tx.outRef<LinearState>(0))
             }
@@ -283,7 +284,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
 
         database.transaction {
             // A tx that spends our money.
-            val spendTXBuilder = TransactionType.General.Builder(DUMMY_NOTARY)
+            val spendTXBuilder = TransactionBuilder(DUMMY_NOTARY)
             vault.generateSpend(spendTXBuilder, 80.DOLLARS, BOB)
             val spendPTX = notaryServices.signInitialTransaction(spendTXBuilder)
             val spendTX = services.addSignature(spendPTX, freshKey)
@@ -313,7 +314,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             linearStates.forEach { println(it.state.data.linearId) }
 
             // Create a txn consuming different contract types
-            val dummyMoveBuilder = TransactionType.General.Builder(notary = DUMMY_NOTARY).apply {
+            val dummyMoveBuilder = TransactionBuilder(notary = DUMMY_NOTARY).apply {
                 addOutputState(DummyLinearContract.State(participants = listOf(freshIdentity)))
                 addOutputState(DummyDealContract.State(ref = "999", participants = listOf(freshIdentity)))
                 addInputState(linearStates.first())